edgebadge: general cleanup, move buttons to its own module
