+++ boards/edgebadge/src/pins.rs
@@ -1,42 +1,35 @@
-//! PyGamer pins
+//! EdgeBadge pins
 
-use super::{hal, pac, pac::MCLK, target_device};
-
-use crate::hal::gpio::{self, *};
-use gpio::{Floating, Input, Output, Port, PushPull};
-use hal::define_pins;
-
-use hal::prelude::*;
-
-use hal::sercom::{
-    I2CMaster2, PadPin, SPIMaster1, SPIMaster4, Sercom2Pad0, Sercom2Pad1, Sercom4Pad1, Sercom4Pad2,
-    Sercom4Pad3, UART5,
-};
+use super::{hal, pac, target_device};
 
 use embedded_hal::{digital::v1_compat::OldOutputPin, timer::CountDown, timer::Periodic};
+use gpio::{Floating, Input, Output, Port, PushPull};
+use hal::clock::GenericClockController;
+use hal::define_pins;
+use hal::gpio::{self, *};
+use hal::hal::spi;
+use hal::prelude::*;
+use hal::sercom::{I2CMaster2, PadPin, SPIMaster1, SPIMaster4, UART5};
+use hal::time::Hertz;
+use pac::MCLK;
+use st7735_lcd::{Orientation, ST7735};
 use ws2812_timer_delay as ws2812;
 
-use hal::clock::GenericClockController;
-
-#[cfg(feature = "unproven")]
-use hal::pwm::Pwm2;
-
-use hal::time::Hertz;
-
-use st7735_lcd::{Orientation, ST7735};
-
-#[cfg(feature = "usb")]
-use super::pac::gclk::{genctrl::SRC_A, pchctrl::GEN_A};
 #[cfg(feature = "usb")]
 use hal::usb::usb_device::bus::UsbBusAllocator;
 #[cfg(feature = "usb")]
 pub use hal::usb::UsbBus;
+#[cfg(feature = "usb")]
+use pac::gclk::{genctrl::SRC_A, pchctrl::GEN_A};
 
 #[cfg(feature = "unproven")]
-use cortex_m::asm::delay as cycle_delay;
-
+pub use crate::buttons::ButtonReader;
 #[cfg(feature = "unproven")]
-use super::pac::ADC0;
+pub use crate::buttons::Keys;
+#[cfg(feature = "unproven")]
+use hal::pwm::Pwm2;
+#[cfg(feature = "unproven")]
+use pac::ADC0;
 
 define_pins!(
     /// Maps the pins to their arduino names and
@@ -326,14 +319,14 @@ impl Display {
         (
             ST7735<
                 SPIMaster4<
-                    Sercom4Pad2<gpio::Pb14<gpio::PfC>>,
-                    Sercom4Pad3<gpio::Pb15<gpio::PfC>>,
-                    Sercom4Pad1<gpio::Pb13<gpio::PfC>>,
+                    hal::sercom::Sercom4Pad2<Pb14<PfC>>,
+                    hal::sercom::Sercom4Pad3<Pb15<PfC>>,
+                    hal::sercom::Sercom4Pad1<Pb13<PfC>>,
                 >,
-                gpio::Pb5<Output<PushPull>>,
-                gpio::Pa0<Output<PushPull>>,
+                Pb5<Output<PushPull>>,
+                Pa0<Output<PushPull>>,
             >,
-            atsamd_hal::samd51::pwm::Pwm2,
+            Pwm2,
         ),
         (),
     > {
@@ -341,9 +334,9 @@ impl Display {
         let tft_spi = SPIMaster4::new(
             &clocks.sercom4_core(&gclk0).ok_or(())?,
             16.mhz(),
-            hal::hal::spi::Mode {
-                phase: hal::hal::spi::Phase::CaptureOnFirstTransition,
-                polarity: hal::hal::spi::Polarity::IdleLow,
+            spi::Mode {
+                phase: spi::Phase::CaptureOnFirstTransition,
+                polarity: spi::Polarity::IdleLow,
             },
             sercom4,
             mclk,
@@ -392,14 +385,7 @@ impl Neopixel {
         self,
         timer: T,
         port: &mut Port,
-    ) -> ws2812::Ws2812<
-        T,
-        embedded_hal::digital::v1_compat::OldOutputPin<
-            atsamd_hal::common::gpio::Pa15<
-                atsamd_hal::common::gpio::Output<atsamd_hal::common::gpio::PushPull>,
-            >,
-        >,
-    > {
+    ) -> ws2812::Ws2812<T, OldOutputPin<Pa15<Output<PushPull>>>> {
         let neopixel_pin: OldOutputPin<_> = self.neopixel.into_push_pull_output(port).into();
 
         ws2812::Ws2812::new(timer, neopixel_pin)
@@ -408,9 +394,9 @@ impl Neopixel {
 
 /// SPI pins
 pub struct SPI {
-    pub mosi: gpio::Pb23<Input<Floating>>,
-    pub miso: gpio::Pb22<Input<Floating>>,
-    pub sck: gpio::Pa17<Input<Floating>>,
+    pub mosi: Pb23<Input<Floating>>,
+    pub miso: Pb22<Input<Floating>>,
+    pub sck: Pa17<Input<Floating>>,
 }
 
 impl SPI {
@@ -424,17 +410,17 @@ impl SPI {
         mclk: &mut MCLK,
         port: &mut Port,
     ) -> SPIMaster1<
-        hal::sercom::Sercom1Pad2<gpio::Pb22<gpio::PfC>>,
-        hal::sercom::Sercom1Pad3<gpio::Pb23<gpio::PfC>>,
-        hal::sercom::Sercom1Pad1<gpio::Pa17<gpio::PfC>>,
+        hal::sercom::Sercom1Pad2<Pb22<PfC>>,
+        hal::sercom::Sercom1Pad3<Pb23<PfC>>,
+        hal::sercom::Sercom1Pad1<Pa17<PfC>>,
     > {
         let gclk0 = clocks.gclk0();
         SPIMaster1::new(
             &clocks.sercom1_core(&gclk0).unwrap(),
             bus_speed.into(),
-            hal::hal::spi::Mode {
-                phase: hal::hal::spi::Phase::CaptureOnFirstTransition,
-                polarity: hal::hal::spi::Polarity::IdleLow,
+            spi::Mode {
+                phase: spi::Phase::CaptureOnFirstTransition,
+                polarity: spi::Polarity::IdleLow,
             },
             sercom1,
             mclk,
@@ -463,7 +449,7 @@ impl I2C {
         sercom2: pac::SERCOM2,
         mclk: &mut MCLK,
         port: &mut Port,
-    ) -> I2CMaster2<Sercom2Pad0<Pa12<PfC>>, Sercom2Pad1<Pa13<PfC>>> {
+    ) -> I2CMaster2<hal::sercom::Sercom2Pad0<Pa12<PfC>>, hal::sercom::Sercom2Pad1<Pa13<PfC>>> {
         let gclk0 = clocks.gclk0();
         I2CMaster2::new(
             &clocks.sercom2_core(&gclk0).unwrap(),
@@ -494,7 +480,7 @@ impl USB {
     /// as a USB device.
     pub fn init(
         self,
-        usb: super::pac::USB,
+        usb: pac::USB,
         clocks: &mut GenericClockController,
         mclk: &mut MCLK,
         port: &mut Port,
@@ -529,12 +515,8 @@ impl UART {
         sercom5: pac::SERCOM5,
         mclk: &mut MCLK,
         port: &mut Port,
-    ) -> UART5<
-        hal::sercom::Sercom5Pad1<gpio::Pb17<gpio::PfC>>,
-        hal::sercom::Sercom5Pad0<gpio::Pb16<gpio::PfC>>,
-        (),
-        (),
-    > {
+    ) -> UART5<hal::sercom::Sercom5Pad1<Pb17<PfC>>, hal::sercom::Sercom5Pad0<Pb16<PfC>>, (), ()>
+    {
         let gclk0 = clocks.gclk0();
 
         UART5::new(
@@ -592,202 +574,6 @@ pub struct Buttons {
 }
 
 #[cfg(feature = "unproven")]
-#[derive(Debug, PartialEq)]
-pub enum Keys {
-    SelectDown,
-    SelectUp,
-    StartDown,
-    StartUp,
-    BDown,
-    BUp,
-    ADown,
-    AUp,
-
-    UpUp,
-    UpDown,
-    DownUp,
-    DownDown,
-    LeftUp,
-    LeftDown,
-    RightUp,
-    RightDown,
-}
-
-#[cfg(feature = "unproven")]
-pub struct ButtonIter {
-    pub pressed: u8,
-    pub released: u8,
-    pub bit_index: u8,
-}
-
-#[cfg(feature = "unproven")]
-//should be impossible for released and pressed, but gives released preference
-fn mask_to_event(mask: u8, released: u8, pressed: u8) -> Option<Keys> {
-    let pressed_bool = mask & pressed == mask;
-    let released_bool = mask & released == mask;
-
-    match mask {
-        0x80 => {
-            if released_bool {
-                Some(Keys::BUp)
-            } else if pressed_bool {
-                Some(Keys::BDown)
-            } else {
-                None
-            }
-        }
-        0x40 => {
-            if released_bool {
-                Some(Keys::AUp)
-            } else if pressed_bool {
-                Some(Keys::ADown)
-            } else {
-                None
-            }
-        }
-        0x20 => {
-            if released_bool {
-                Some(Keys::StartUp)
-            } else if pressed_bool {
-                Some(Keys::StartDown)
-            } else {
-                None
-            }
-        }
-        0x10 => {
-            if released_bool {
-                Some(Keys::SelectUp)
-            } else if pressed_bool {
-                Some(Keys::SelectDown)
-            } else {
-                None
-            }
-        }
-
-        0x08 => {
-            if released_bool {
-                Some(Keys::RightUp)
-            } else if pressed_bool {
-                Some(Keys::RightDown)
-            } else {
-                None
-            }
-        }
-        0x04 => {
-            if released_bool {
-                Some(Keys::DownUp)
-            } else if pressed_bool {
-                Some(Keys::DownDown)
-            } else {
-                None
-            }
-        }
-        0x02 => {
-            if released_bool {
-                Some(Keys::UpUp)
-            } else if pressed_bool {
-                Some(Keys::UpDown)
-            } else {
-                None
-            }
-        }
-        0x01 => {
-            if released_bool {
-                Some(Keys::LeftUp)
-            } else if pressed_bool {
-                Some(Keys::LeftDown)
-            } else {
-                None
-            }
-        }
-        _ => None,
-    }
-}
-
-#[cfg(feature = "unproven")]
-impl Iterator for ButtonIter {
-    type Item = Keys;
-
-    fn next(&mut self) -> Option<Keys> {
-        //really want a while post increment but doesnt exist
-        if self.bit_index >= 8 {
-            return None;
-        }
-
-        //funky do while
-        while {
-            let mask = 0x01 << self.bit_index;
-            self.bit_index += 1;
-
-            let event = mask_to_event(mask, self.released, self.pressed);
-            if event.is_some() {
-                return event;
-            }
-
-            self.bit_index < 8
-        } {}
-
-        None
-    }
-}
-
-/// Button pins
-#[cfg(feature = "unproven")]
-pub struct ButtonReader {
-    /// Button Latch
-    pub latch: Pb0<Output<PushPull>>,
-    /// Button Out
-    pub data_in: Pb30<Input<Floating>>,
-    /// Button Clock
-    pub clock: Pb31<Output<PushPull>>,
-    pub last: u8,
-}
-
-//120mhz, 1 cycle = 0.000000008333333 = 8.333333ns
-//https://www.onsemi.com/pub/Collateral/MC74HC165A-D.PDF
-//3v <=125c
-//tsu min setup time 55ns = 7 cycles
-//th min hold time 5ns = 1 cycles
-//tw min pulse width 36ns = 5 cycles
-//trec min recovery time 55ns, how long before you should attempt to read
-// again?
-#[cfg(feature = "unproven")]
-impl ButtonReader {
-    // 48*8.333ns total blocking read
-    /// Returns a ButtonIter of button changes as Keys enums
-    pub fn events(&mut self) -> ButtonIter {
-        self.latch.set_low().ok();
-        cycle_delay(7); //tsu?
-        self.latch.set_high().ok();
-        cycle_delay(1); //th?
-
-        let mut current: u8 = 0;
-
-        for _i in 0..8 {
-            current <<= 1;
-
-            self.clock.set_low().ok();
-            cycle_delay(5); //tw
-
-            if self.data_in.is_high().unwrap() {
-                current |= 1;
-            }
-            self.clock.set_high().ok();
-        }
-
-        let iter = ButtonIter {
-            pressed: (self.last ^ current) & current,
-            released: (self.last ^ current) & self.last,
-            bit_index: 0,
-        };
-
-        self.last = current;
-
-        iter
-    }
-}
-
-#[cfg(feature = "unproven")]
 impl Buttons {
     /// Convenience for setting up the button latch pins
     /// Returns ButtonReader iterator which can be polled for Key events
@@ -813,7 +599,7 @@ impl Buttons {
 #[cfg(feature = "unproven")]
 pub struct BatteryReader {
     /// Battery pin
-    pub battery: gpio::Pb1<gpio::PfB>,
+    pub battery: Pb1<PfB>,
 }
 
 #[cfg(feature = "unproven")]
@@ -839,101 +625,5 @@ impl Battery {
         BatteryReader {
             battery: self.battery.into_function_b(port),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_from_left() {
-        let mut iter = ButtonIter {
-            pressed: 0x0F,
-            released: 0x00,
-            bit_index: 0,
-        };
-
-        assert_eq!(Some(Keys::BDown), iter.next());
-        assert_eq!(Some(Keys::ADown), iter.next());
-        assert_eq!(Some(Keys::StartDown), iter.next());
-        assert_eq!(Some(Keys::SelectDown), iter.next());
-        assert_eq!(None, iter.next());
-    }
-
-    #[test]
-    fn iter_released_before_pressed() {
-        let mut iter = ButtonIter {
-            pressed: 0x0F,
-            released: 0x0F,
-            bit_index: 0,
-        };
-
-        assert_eq!(Some(Keys::BUp), iter.next());
-        assert_eq!(Some(Keys::AUp), iter.next());
-        assert_eq!(Some(Keys::StartUp), iter.next());
-        assert_eq!(Some(Keys::SelectUp), iter.next());
-        assert_eq!(None, iter.next());
-    }
-
-    #[test]
-    fn iter_released_before_pressed2() {
-        let mut iter = ButtonIter {
-            pressed: 0x05,
-            released: 0x0A,
-            bit_index: 0,
-        };
-
-        assert_eq!(Some(Keys::BUp), iter.next());
-        assert_eq!(Some(Keys::ADown), iter.next());
-        assert_eq!(Some(Keys::StartUp), iter.next());
-        assert_eq!(Some(Keys::SelectDown), iter.next());
-        assert_eq!(None, iter.next());
-    }
-
-    #[test]
-    fn inner_empty_bit_is_skipped() {
-        let mut iter = ButtonIter {
-            pressed: 0x05,
-            released: 0x00,
-            bit_index: 0,
-        };
-
-        assert_eq!(Some(Keys::ADown), iter.next());
-        assert_eq!(Some(Keys::SelectDown), iter.next());
-        assert_eq!(None, iter.next());
-    }
-
-    #[test]
-    fn empty_returns_none() {
-        let mut iter = ButtonIter {
-            pressed: 0x00,
-            released: 0x00,
-            bit_index: 0,
-        };
-
-        assert_eq!(None, iter.next());
-    }
-
-    #[test]
-    fn upper_four_bits_unused() {
-        let mut iter = ButtonIter {
-            pressed: 0xF0,
-            released: 0x00,
-            bit_index: 0,
-        };
-
-        assert_eq!(None, iter.next());
-    }
-
-    #[test]
-    fn upper_four_bits_unused2() {
-        let mut iter = ButtonIter {
-            pressed: 0x00,
-            released: 0xF0,
-            bit_index: 0,
-        };
-
-        assert_eq!(None, iter.next());
     }
 }
